modify Log model and merge 'save_input' to 'input' url
modify Log hours fields to be able to input float and do so in form
checking Log model before save model in input view, change logic there
is only one Log model how many user input log

+++ main/forms.py
@@ -7,7 +7,7 @@ class CreateLog(forms.ModelForm):
     date = forms.DateField(initial=timezone.now)
     start = forms.CharField(initial="00:00")
     end = forms.CharField(initial="00:00")
-    hours = forms.IntegerField(initial=0)
+    hours = forms.FloatField(initial=0)
 
     class Meta:
         model = Log
